% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHEhmc-package.R
\docType{data}
\name{internal_stan}
\alias{internal_stan}
\alias{stanmodels}
\alias{stan_files}
\alias{model_Exponential}
\alias{model_Gamma}
\alias{model_GenF}
\alias{model_GenGamma}
\alias{model_Gompertz}
\alias{model_logLogistic}
\alias{model_logNormal}
\alias{model_PolyWeibull}
\alias{model_RP}
\alias{model_WeibullAF}
\alias{model_WeibullPH}
\alias{Rcpp_model_Exponential-class}
\alias{Rcpp_model_Gamma-class}
\alias{Rcpp_model_GenF-class}
\alias{Rcpp_model_GenGamma-class}
\alias{Rcpp_model_Gompertz-class}
\alias{Rcpp_model_logLogistic-class}
\alias{Rcpp_model_logNormal-class}
\alias{Rcpp_model_PolyWeibull-class}
\alias{Rcpp_model_RP-class}
\alias{Rcpp_model_WeibullAF-class}
\alias{Rcpp_model_WeibullPH-class}
\title{Internal objects used by stan}
\description{
Objects, C++ and S4 classes used by stan to fit parametric survival models.
}
\keyword{internal}
