% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.transition.probs.R
\name{make.transition.probs}
\alias{make.transition.probs}
\title{Compute transition probabilities}
\usage{
make.transition.probs(x, ...)
}
\arguments{
\item{x}{an object obtained as output of the call to \code{ \link{make.surv}}}

\item{...}{additional arguments. Includes labs = a string vector of names for the elements of 
the list (strata for the survival analysis)}
}
\value{
list of matrices with \code{nsim} rows and \code{length(time)} columns
with the simulations for the transition probabilities between consecutive times.
}
\description{
Computes the transition probabilities (to be passed to a Markov model) from the survival
curves obtained using \code{fit.models} and \code{make.surv}, using the formula 
p(t)=1-S(t+k)/S(t), where k is the Markov model cycle length and t is a generic time
}
\author{
Gianluca Baio
}
