% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.transition.probs.R
\name{markov_trace}
\alias{markov_trace}
\title{markov_trace}
\usage{
markov_trace(mm, interventions = NULL, ...)
}
\arguments{
\item{mm}{}

\item{...}{additional arguments.}
}
\value{

}
\description{
Computes the transition probabilities (to be passed to a Markov model) from
the survival curves obtained using \code{fit.models}, using the formula 
p(t)=1-exp(H(t+k)-H(t), where k is the Markov model cycle length and t is a 
generic time
}
\note{
Something will go here
}
\examples{
\dontrun{
# Something will go here
}

}
\references{
Something will go here
}
\seealso{
make.surv
}
\author{
Gianluca Baio
}
\keyword{Markov}
\keyword{Transition}
\keyword{models}
\keyword{probabilities}
\keyword{trace}
