% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.fit.plot.R
\name{model.fit.plot}
\alias{model.fit.plot}
\title{Graphical representation of the measures of model fitting based on Information Criteria}
\usage{
model.fit.plot(..., type = "aic")
}
\arguments{
\item{...}{Optional inputs. Must include at least one \code{survHE} object.}

\item{type}{should the AIC, BIC or DIC be plotted? (values = \code{"aic"}, \code{"bic"} or \code{"dic"})}
}
\value{
barplot of the selected Information Criteria for each model
}
\description{
Plots a summary of the model fit for all the models fitted
}
\author{
Gianluca Baio
}

