% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survHE.R
\name{plot.survHE}
\alias{plot.survHE}
\title{Plot survival curves for the fitted models}
\usage{
\method{plot}{survHE}(...)
}
\arguments{
\item{...}{Must include at least one result object saved as the call to the \code{fit.models} function.
See details for other options.}
}
\value{
a plot containing the Kaplan-Meier curve and the fitted model
}
\description{
Plots the results of model fit.
}
\details{
Other possibilities are additional (mainly graphical) options. These are:
\itemize{
\item{xlab} {a string with the label for the x-axis (default = "time")}
\item{ylab} {a string with the label for the y-axis (default = "Survival")}
\item{lab.trt} {a (vector of) string(s) indicating the labels associated with the
strata defining the different survival curves to plot. Default to the value used by
the Kaplan Meier estimate given in \code{\link{fit.models}}}
\item{cex.trt} {factor by which the size of the font used to write the strata is 
resized (default = 0.8)}
\item{n.risk} {logical. If \code{TRUE} (default) writes the number at risk at different
time points (as determined by the Kaplan Meier estimate)}
\item{newdata} {a list (of lists) providing the values for the relevant covariates
If NULL, then will use the mean values for the covariates if at least one is 
a continuous variable, or the combination of the categorical covariates.}
\item{xlim} {a vector determining the limits for the x-axis}
\item{colors} {a vector of characters defining the colours in which to plot the
different survival curves}
\item{labs} {a vector of characters defining the names of the models fitted}
\item{add.km} {whether to also add the Kaplan Meier estimates of the data. Defaults to \code{TRUE}}
}
}
\examples{
data(bc)
mle <- fit.models(formula = Surv(recyrs, censrec) ~ group, data = bc,
   distr = "exp", method = "mle")
plot(mle)
}
\author{
Gianluca Baio
}
