% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survHE.R
\name{print.survHE}
\alias{print.survHE}
\title{Print a summary of the survival model(s)}
\usage{
\method{print}{survHE}(x, mod = 1, ...)
}
\arguments{
\item{x}{the \code{survHE} object (the output of the call to \code{fit.models})}

\item{mod}{the index of the model. Default value is 1, but the user can choose which model 
fit to visualise, if the call to fit.models has a vector argument for distr (so many 
models are fitted & stored in the same object)}

\item{...}{additional options, including:
\itemize{
\item{digits} {number of significant digits to be shown in the summary table 
(default = 6)}
\item{nsim} {number of simulations from the joint posterior for \code{INLA} 
(default = 100)}
\item{original} {a flag to say whether the original table from either \code{flexsurv}
or \code{INLA} or \code{rstan} should be printed}
}}
}
\value{
prints a summary table
}
\description{
Prints the summary table for the model(s) fitted, with the estimate of the parameters
}
\examples{
data(bc)
mle = fit.models(formula = Surv(recyrs, censrec) ~ group, data = bc,
    distr = "exp", method = "mle")
print(mle)
}
\author{
Gianluca Baio
}

