% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHMC.R
\name{runHMC}
\alias{runHMC}
\title{for a given formula and dataset}
\usage{
runHMC(x, exArgs)
}
\arguments{
\item{x}{a (vector of) string(s) containing the name(s) of the model(s)
to be fitted}

\item{exArgs}{a list of extra arguments passed from the main 'fit.models'
function}
}
\description{
for a given formula and dataset
}
\note{
Something will go here
}
\references{
Baio (2020). survHE
}
\seealso{
fit.models
}
\author{
Gianluca Baio
}
\keyword{Carlo}
\keyword{Hamiltonian}
\keyword{Monte}
\keyword{Parametric}
\keyword{models}
\keyword{survival}
