% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runINLA.R
\name{runINLA}
\alias{runINLA}
\title{Helper function to run the survival models using INLA
for a given formula and dataset}
\usage{
runINLA(x, exArgs)
}
\arguments{
\item{x}{a (vector of) string(s) containing the name(s) of the model(s)
to be fitted}

\item{exArgs}{a list of extra arguments passed from the main 'fit.models'
function}
}
\description{
Helper function to run the survival models using INLA
for a given formula and dataset
}
\note{
Something will go here
}
\references{
Baio (2020). survHE
}
\seealso{
fit.models
}
\author{
Gianluca Baio
}
\keyword{Approximation}
\keyword{Integrated}
\keyword{Laplace}
\keyword{Nested}
\keyword{Parametric}
\keyword{models}
\keyword{survival}
