% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHE-package.R
\docType{package}
\name{survHE}
\alias{survHE}
\alias{survHE-package}
\title{survHE: Survival Analysis in Health Economic Evaluation}
\description{
Contains a suite of functions to systematise the workflow involving survival
analysis in health economic evaluation. survHE can fit a large range of
survival models using both a frequentist approach (by calling the R package
flexsurv and a Bayesian perspective.
}

