% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ta174}
\alias{ta174}
\title{NICE TA174 dataset.}
\format{
A tibble with 810 rows and 8 variables:
\describe{
\item{patid}{A numeric patient identifier}
\item{treat}{The treatment indicator. 1=rituximab
in combination with fludarabine andcyclophosphamide
(RFC); 0=fludarabine and cyclo-phosphamide alone (FC)}
\item{prog}{An indicator to describe whether the
patient has experience a progression}
\item{death}{An indicator to describe whether the
patient has experience death}
\item{prog_t}{The observed time at progression, or
the time at which the patient has been censored;
measured in months}
\item{death_t}{The observed time at death, or
the time at which the patient has been censored;
measured in months}
\item{prog_ty}{The observed time at progression, or
the time at which the patient has been censored;
measured in years}
\item{death_ty}{The observed time at death, or
the time at which the patient has been censored;
measured in years}
}
}
\usage{
ta174
}
\description{
A dataset containing the data used for NICE TA174, as
made publicly available as part of the supplementary
material for Williams et al (2017). Medical Decision
Making, 37;427-439.
}
\keyword{datasets}
