% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.surv.R
\name{write.surv}
\alias{write.surv}
\title{Writes the survival summary to an excel file}
\usage{
write.surv(object, file, sheet = NULL, what = "surv")
}
\arguments{
\item{object}{a \code{summary.flexsurvreg} object containing the survival curves (with times, estimates and 
interval limits)}

\item{file}{a string with the full path to the file name to be saved}

\item{sheet}{a string with the name of the sheet to be created}

\item{what}{a string to describe what to be exported. Possible values are:
\describe{
\item{surv}{a matrix with nsim rows and ntimes columns with the survival
curve (one such matix for each configuration of the covariates)}
\item{sim}{a matrix with nsim rows and simulations for the survival parameters (scale, shape, rate, etc)}
}}
}
\value{
an excel file
}
\description{
Writes the survival summary to an excel file (helpful to then call the values 
in the Markov model)
}
\author{
Gianluca Baio
}

