% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.transition.probs.R
\name{make.transition.probs}
\alias{make.transition.probs}
\title{make.transition.probs}
\usage{
make.transition.probs(x, ...)
}
\arguments{
\item{x}{an object obtained as output of the call to \code{make.surv}}

\item{...}{additional arguments. Includes labs = a string vector of names
for the elements of the list (strata for the survival analysis)}
}
\value{
Something will go here
}
\description{
Computes the transition probabilities (to be passed to a Markov model) from
the survival curves obtained using \code{fit.models} and \code{make.surv},
using the formula p(t)=1-S(t+k)/S(t), where k is the Markov model cycle
length and t is a generic time
}
\details{
Something will go here
}
\note{
Something will go here
}
\examples{
\dontrun{
# Something will go here
}

}
\references{
Something will go here
}
\seealso{
Something will go here
}
\author{
Gianluca Baio
}
\keyword{Markov}
\keyword{Transition}
\keyword{models}
\keyword{probabilities}
