% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.transition.probs.R
\name{markov_trace}
\alias{markov_trace}
\title{Markov trace}
\usage{
markov_trace(mm, interventions = NULL, ...)
}
\arguments{
\item{mm}{The output of a call to \code{three_state_mm}}

\item{interventions}{A vector of labels for the interventions}

\item{...}{additional arguments.}
}
\value{
Plot
}
\description{
Plots the Markov Trace from an object generated using \code{three_state_mm}
}
\note{
Something will go here
}
\examples{
\dontrun{
# Something will go here
}

}
\references{
Something will go here
}
\seealso{
make.surv, three_state_mm
}
\author{
Gianluca Baio
}
\keyword{Markov}
\keyword{Transition}
\keyword{models}
\keyword{probabilities}
\keyword{trace}
