% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.fit.plot.R
\name{model.fit.plot}
\alias{model.fit.plot}
\title{Graphical representation of the measures of model fitting based on
Information Criteria}
\usage{
model.fit.plot(..., type = "aic", scale = "absolute")
}
\arguments{
\item{...}{Optional inputs. Must include at least one \code{survHE} object.}

\item{type}{should the AIC, the BIC or the DIC plotted? (values = \code{"aic"},
\code{"bic"} or \code{"dic"})}

\item{scale}{If \code{scale='absolute'} (default), then plot the absolute value 
of the *IC. If \code{scale='relative'} then plot a rescaled version taking
the percentage increase in the *IC in comparison with the best-fitting model}
}
\value{
Something will go here
}
\description{
Plots a summary of the model fit for all the models fitted
}
\details{
Something will go here
}
\note{
Something will go here
}
\examples{

# Something will go here

}
\references{
Something will go here
}
\seealso{
Something will go here
}
\author{
Gianluca Baio
}
\keyword{Model}
\keyword{Parametric}
\keyword{fitting}
\keyword{models}
\keyword{survival}
