\name{model.fit.plot}
\alias{model.fit.plot}
\title{
Graphical representation of the measures of model fitting based on Information Criteria
}
\description{
Plots a summary of the model fit for all the models fitted
}
\usage{
model.fit.plot(..., type = "aic")
}
\arguments{
\item{...}{
Optional inputs. Must include at least one \code{survHE} object.
}
\item{type}{
should the AIC or the BIC plotted? (values = \code{"aic"}, \code{"bic"} or \code{"dic"})
}
}
\details{
Something will go here
}
\value{
Something will go here
}
\references{
Something will go here
}
\author{
Gianluca Baio
}
\note{
Something will go here
}
\seealso{
Something will go here
}
\examples{
# Something will go here
}
\keyword{Model fitting}
\keyword{Parametric survival models}% __ONLY ONE__ keyword per line
