% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHE-package.R
\docType{package}
\name{survHE-package}
\alias{survHE-package}
\alias{survHE}
\title{Survival Analysis in Health Economic Evaluation}
\description{
Contains a suite of functions to perform survival analysis with the aim of
aiding in health economic modelling (extrapolation, model checking and PSA)
}
\details{
\tabular{ll}{ Package: \tab survHE\cr Type: \tab Package\cr Version: \tab
1.0.66\cr Date: \tab 2019-09-12\cr License: \tab GPL2 \cr LazyLoad: \tab
yes\cr } Contains a suite of functions to perform survival analysis with the
aim of aiding in health economic modelling (extrapolation, model checking
and PSA)
}
\examples{

# Something will go here

}
\references{
Something will go here
}
\seealso{
Something will go here
}
\author{
Gianluca Baio

Maintainer: Gianluca Baio
}
