\name{survHE-package}
\alias{survHE-package}
\alias{survHE}
\docType{package}
\title{
Survival Analysis in Health Economic Evaluation
}
\description{
Contains a suite of functions to perform survival analysis with the aim of aiding
in health economic modelling (extrapolation, model checking and PSA)
}
\details{
\tabular{ll}{
Package: \tab survHE\cr
Type: \tab Package\cr
Version: \tab 1.0.4\cr
Date: \tab 2016-12-04\cr
License: \tab GPL2 \cr
LazyLoad: \tab yes\cr
}
Contains a suite of functions to perform survival analysis with the aim of aiding
in health economic modelling (extrapolation, model checking and PSA)
}
\author{
Gianluca Baio

Maintainer: Gianluca Baio
}
\references{
Something will go here
}
\keyword{
Something will go here
}
\seealso{
Something will go here
}
\examples{
# Something will go here
}
