% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.linear.assumptions.R
\name{test.linear.assumptions}
\alias{test.linear.assumptions}
\title{Tests the linear assumptions for the parametric model}
\usage{
test.linear.assumptions(fit, mod = 1, label_plot = FALSE, ...)
}
\arguments{
\item{fit}{an object of class survHE}

\item{mod}{index or name of a model in fit. Defaults to 1.}

\item{label_plot}{if TRUE, labels assumptions. Defaults to FALSE.}

\item{\dots}{further arguments, passed on to points()}
}
\value{
A diagnostic plot
}
\description{
Tests the linear assumptions for the parametric model
}
\author{
William Browne
}
\keyword{hplot}
\keyword{survival}
