% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.surv.R
\name{write.surv}
\alias{write.surv}
\title{write.surv}
\usage{
write.surv(object, file, sheet = NULL, what = "surv")
}
\arguments{
\item{object}{a summary.flexsurvreg object containing the survival curves
(with times, estimates and interval limits)}

\item{file}{a string with the full path to the file name to be saved}

\item{sheet}{a string with the name of the sheet to be created}

\item{what}{a string to describe what to be exported}
}
\value{
Something will go here
}
\description{
Writes the survival summary to an excel file (helpful to then call the
values in the Markov model)
}
\details{
Something will go here
}
\note{
Something will go here
}
\examples{

# Something will go here

}
\references{
Something will go here
}
\seealso{
Something will go here
}
\author{
Gianluca Baio
}
\keyword{Excel}
